refactor: pageなのでディレクトリを作成して移動
